FEAT: Proper Error impls for errors
Also use pub(crate), requiring Rust 1.18
